Refactoring, Implemented Duplicate Prevention, Tests
Refactored a little bit in the 15-puzzle-game and bot file to make code more readable and remove useless functions. Implemented a hardcoded test case in botTest file, which oddly enough works perfectly. Also implemented logic to not allow for duplicate states to be visited, which optimally wouldn't happen but thats besides the point. Praying this eventually works somehow.